Imported Scriptable Event System
To use the event system, serialize a GameEvent variable for the broadcaster, create a GameEvent scriptable object for the new "event channel", drag the new SO (scriptable object) into the serialized field. Whatever object should listen for the event should have an EventListener component on it, in the inspector, link the unity event to whatever method you would like to invoke.